Refactor VM modules.
