+++ crates/motoko/src/lib/mod.rs
@@ -26,9 +26,13 @@ pub mod proc_macro;
 mod serde_utils;
 pub mod shared;
 pub mod value;
+
 pub mod vm;
 pub mod vm_core;
+pub mod vm_def;
+pub mod vm_match;
 pub mod vm_ops;
+pub mod vm_step;
 pub mod vm_types;
 
 #[cfg(feature = "parser")]
