+++ crates/motoko/src/lib/vm_core.rs
@@ -1,20 +1,26 @@
+use crate::ast::PrimType;
 use crate::ast::{Dec, DecField, Dec_, Exp_, Id, Id_, Inst, Prog, Source, ToId};
-//use crate::ast_traversal::ToNode;
 use crate::shared::{FastClone, Share};
 use crate::value::{ActorId, ActorMethod, Value, Value_};
+use crate::vm_def::def;
+use crate::vm_types::def::CtxId;
+use crate::vm_types::ActiveBorrow;
+use crate::vm_types::Actor;
+use crate::vm_types::DebugPrintLine;
+use crate::vm_types::Env;
+use crate::vm_types::ModuleFileInit;
+use crate::vm_types::Stack;
 use crate::vm_types::{
-    def::{Def, Defs, Field as FieldDef},
+    def::{Actor as ActorDef, Def, Defs, Field as FieldDef, Module as ModuleDef},
     stack::{Frame, FrameCont},
     Activation, Active, Actors, Agent, Cont, Core, Counts, Interruption, Limits, ModuleFileState,
     ModuleFiles, ModulePath, Pointer, Response, ScheduleChoice, Step, SyntaxError,
 };
 use crate::vm_types::{EvalInitError, Store};
+use crate::vm_types::{LocalPointer, NamedPointer};
+use crate::{nyi, type_mismatch};
 use im_rc::{HashMap, Vector};
 use std::vec::Vec;
-
-use crate::type_mismatch;
-use crate::vm::def;
-use crate::vm_types::ModuleFileInit;
 
 fn agent_init(prog: Prog) -> Agent {
     let mut a = Agent {
@@ -25,6 +31,392 @@ fn agent_init(prog: Prog) -> Agent {
     };
     a.active.cont = Cont::Decs(prog.vec);
     a
+}
+
+impl Active for Core {
+    fn ctx_id<'a>(&'a mut self) -> &'a mut CtxId {
+        &mut self.defs.active_ctx
+    }
+    fn defs<'a>(&'a mut self) -> &'a mut Defs {
+        &mut self.defs
+    }
+    fn module_files<'a>(&'a mut self) -> &'a mut ModuleFiles {
+        &mut self.module_files
+    }
+    //fn schedule_choice<'a>(&'a self) -> &'a ScheduleChoice {
+    //&self.schedule_choice
+    //}
+    fn cont<'a>(&'a mut self) -> &'a mut Cont {
+        use ScheduleChoice::*;
+        match &self.schedule_choice {
+            Agent => &mut self.agent.active.cont,
+            Actor(ref n) => {
+                &mut self
+                    .actors
+                    .map
+                    .get_mut(n)
+                    .unwrap()
+                    .active
+                    .as_mut()
+                    .unwrap()
+                    .cont
+            }
+        }
+    }
+    fn package<'a>(&'a mut self) -> &'a mut Option<String> {
+        use ScheduleChoice::*;
+        match &self.schedule_choice {
+            Agent => &mut self.agent.active.package,
+            Actor(ref n) => {
+                &mut self
+                    .actors
+                    .map
+                    .get_mut(n)
+                    .unwrap()
+                    .active
+                    .as_mut()
+                    .unwrap()
+                    .package
+            }
+        }
+    }
+    fn cont_source<'a>(&'a mut self) -> &'a mut Source {
+        use ScheduleChoice::*;
+        match &self.schedule_choice {
+            Agent => &mut self.agent.active.cont_source,
+            Actor(ref n) => {
+                &mut self
+                    .actors
+                    .map
+                    .get_mut(n)
+                    .unwrap()
+                    .active
+                    .as_mut()
+                    .unwrap()
+                    .cont_source
+            }
+        }
+    }
+    fn cont_prim_type<'a>(&'a mut self) -> &'a mut Option<PrimType> {
+        use ScheduleChoice::*;
+        match &self.schedule_choice {
+            Agent => &mut self.agent.active.cont_prim_type,
+            Actor(ref n) => {
+                &mut self
+                    .actors
+                    .map
+                    .get_mut(n)
+                    .unwrap()
+                    .active
+                    .as_mut()
+                    .unwrap()
+                    .cont_prim_type
+            }
+        }
+    }
+    fn env<'a>(&'a mut self) -> &'a mut Env {
+        use ScheduleChoice::*;
+        match &self.schedule_choice {
+            Agent => &mut self.agent.active.env,
+            Actor(ref n) => {
+                &mut self
+                    .actors
+                    .map
+                    .get_mut(n)
+                    .unwrap()
+                    .active
+                    .as_mut()
+                    .unwrap()
+                    .env
+            }
+        }
+    }
+    fn stack<'a>(&'a mut self) -> &'a mut Stack {
+        use ScheduleChoice::*;
+        match &self.schedule_choice {
+            Agent => &mut self.agent.active.stack,
+            Actor(ref n) => {
+                &mut self
+                    .actors
+                    .map
+                    .get_mut(n)
+                    .unwrap()
+                    .active
+                    .as_mut()
+                    .unwrap()
+                    .stack
+            }
+        }
+    }
+    fn store<'a>(&'a mut self) -> &'a mut Store {
+        use ScheduleChoice::*;
+        match &self.schedule_choice {
+            Agent => &mut self.agent.store,
+            Actor(ref n) => &mut self.actors.map.get_mut(n).unwrap().store,
+        }
+    }
+    fn debug_print_out<'a>(&'a mut self) -> &'a mut Vector<DebugPrintLine> {
+        &mut self.debug_print_out
+    }
+    fn counts<'a>(&'a mut self) -> &'a mut Counts {
+        use ScheduleChoice::*;
+        match &self.schedule_choice {
+            Agent => &mut self.agent.counts,
+            Actor(ref n) => &mut self.actors.map.get_mut(n).unwrap().counts,
+        }
+    }
+
+    fn create(
+        &mut self,
+        path: String,
+        name: ActorId,
+        def: ActorDef,
+    ) -> Result<Value_, Interruption> {
+        let v = Value::Actor(crate::value::Actor {
+            def: Some(def.clone()),
+            id: name.clone(),
+        });
+        //let def = self.defs().map.get(&CtxId(0)).unwrap().fields.get(name).unwrap().def.clone();
+        let mut store = Store::new(ScheduleChoice::Actor(name.clone()));
+        let mut env = self.env().clone();
+        let ctx = self.defs().map.get(&def.fields).unwrap();
+        for (i, field) in ctx.fields.iter() {
+            match &field.def {
+                Def::Var(v) => {
+                    store.alloc_named(i.clone(), v.init.fast_clone());
+                    let owner = ScheduleChoice::Actor(name.clone());
+                    env.insert(
+                        i.clone(),
+                        Value::Pointer(Pointer {
+                            owner,
+                            local: LocalPointer::Named(NamedPointer(i.clone())),
+                        })
+                        .share(),
+                    );
+                }
+                Def::Func(f) => {
+                    env.insert(i.clone(), f.rec_value.fast_clone());
+                }
+                _ => return nyi!(line!()),
+            }
+        }
+        let a = Actor {
+            path,
+            def,
+            env,
+            store,
+            counts: Counts::default(),
+            active: None,
+            awaiting: HashMap::new(),
+        };
+        let a0 = self.actors.map.insert(name, a);
+        if let Some(_a0) = a0 {
+            todo!("upgrade")
+        };
+        Ok(v.share())
+    }
+
+    fn upgrade(
+        &mut self,
+        path: String,
+        name: ActorId,
+        def: ActorDef,
+    ) -> Result<Value_, Interruption> {
+        let v = Value::Actor(crate::value::Actor {
+            def: Some(def.clone()),
+            id: name.clone(),
+        });
+        let mut env = HashMap::new();
+        let mut store = self.actors.map.get(&name).unwrap().store.clone();
+        let counts = self.actors.map.get(&name).unwrap().counts.clone();
+        let ctx = self.defs().map.get(&def.fields).unwrap();
+        for (i, field) in ctx.fields.iter() {
+            match &field.def {
+                Def::Var(v) => {
+                    match store.get(&LocalPointer::Named(NamedPointer(i.clone()))) {
+                        None => {
+                            let p = store.alloc_named(i.clone(), v.init.fast_clone());
+                            let pv = Value::Pointer(p).share();
+                            env.insert(i.clone(), pv);
+                        }
+                        Some(_) => {
+                            let p = Pointer {
+                                owner: ScheduleChoice::Actor(name.clone()),
+                                local: LocalPointer::Named(NamedPointer(i.clone())),
+                            };
+                            let pv = Value::Pointer(p).share();
+                            // keep store's current value.
+                            // (even if not stable.)
+                            env.insert(i.clone(), pv);
+                        }
+                    }
+                }
+                Def::Func(..) => {
+                    // to do
+                }
+                _ => todo!(),
+            }
+        }
+        let a = Actor {
+            path,
+            def,
+            env,
+            store,
+            counts,
+            active: None,
+            awaiting: HashMap::new(),
+        };
+        self.actors.map.insert(name, a);
+        Ok(v.share())
+    }
+
+    fn create_module(
+        &mut self,
+        _path: ModulePath,
+        _id: Option<Id>,
+        module: ModuleDef,
+    ) -> Result<Value_, Interruption> {
+        Ok(crate::value::Value::Module(module).share())
+    }
+
+    fn upgrade_module(
+        &mut self,
+        _path: ModulePath,
+        _id: Option<Id>,
+        module: ModuleDef,
+    ) -> Result<Value_, Interruption> {
+        Ok(crate::value::Value::Module(module).share())
+    }
+}
+
+impl ActiveBorrow for Core {
+    fn ctx_id<'a>(&'a self) -> &'a CtxId {
+        &self.defs.active_ctx
+    }
+    fn defs<'a>(&'a self) -> &'a Defs {
+        &self.defs
+    }
+    fn schedule_choice<'a>(&'a self) -> &'a ScheduleChoice {
+        &self.schedule_choice
+    }
+    fn cont<'a>(&'a self) -> &'a Cont {
+        use ScheduleChoice::*;
+        match &self.schedule_choice {
+            Agent => &self.agent.active.cont,
+            Actor(ref n) => {
+                &self
+                    .actors
+                    .map
+                    .get(n)
+                    .unwrap()
+                    .active
+                    .as_ref()
+                    .unwrap()
+                    .cont
+            }
+        }
+    }
+    fn cont_source<'a>(&'a self) -> &'a Source {
+        use ScheduleChoice::*;
+        match &self.schedule_choice {
+            Agent => &self.agent.active.cont_source,
+            Actor(ref n) => {
+                &self
+                    .actors
+                    .map
+                    .get(n)
+                    .unwrap()
+                    .active
+                    .as_ref()
+                    .unwrap()
+                    .cont_source
+            }
+        }
+    }
+    fn cont_prim_type<'a>(&'a self) -> &'a Option<PrimType> {
+        use ScheduleChoice::*;
+        match &self.schedule_choice {
+            Agent => &self.agent.active.cont_prim_type,
+            Actor(ref n) => {
+                &self
+                    .actors
+                    .map
+                    .get(n)
+                    .unwrap()
+                    .active
+                    .as_ref()
+                    .unwrap()
+                    .cont_prim_type
+            }
+        }
+    }
+    fn env<'a>(&'a self) -> &'a Env {
+        use ScheduleChoice::*;
+        match &self.schedule_choice {
+            Agent => &self.agent.active.env,
+            Actor(ref n) => &self.actors.map.get(n).unwrap().active.as_ref().unwrap().env,
+        }
+    }
+    fn stack<'a>(&'a self) -> &'a Stack {
+        use ScheduleChoice::*;
+        match &self.schedule_choice {
+            Agent => &self.agent.active.stack,
+            Actor(ref n) => {
+                &self
+                    .actors
+                    .map
+                    .get(n)
+                    .unwrap()
+                    .active
+                    .as_ref()
+                    .unwrap()
+                    .stack
+            }
+        }
+    }
+    fn store<'a>(&'a self) -> &'a Store {
+        use ScheduleChoice::*;
+        match &self.schedule_choice {
+            Agent => &self.agent.store,
+            Actor(ref n) => &self.actors.map.get(n).unwrap().store,
+        }
+    }
+    fn debug_print_out<'a>(&'a self) -> &'a Vector<DebugPrintLine> {
+        &self.debug_print_out
+    }
+    fn counts<'a>(&'a self) -> &'a Counts {
+        use ScheduleChoice::*;
+        match &self.schedule_choice {
+            Agent => &self.agent.counts,
+            Actor(ref n) => &self.actors.map.get(n).unwrap().counts,
+        }
+    }
+}
+
+impl Limits {
+    pub fn default() -> Self {
+        Self::none()
+    }
+
+    /// No limits.
+    pub fn none() -> Limits {
+        Limits {
+            breakpoints: vec![],
+            step: None,
+            redex: None,
+            send: None,
+        }
+    }
+    /// Set step limit.
+    pub fn step(mut self, s: usize) -> Self {
+        self.step = Some(s);
+        self
+    }
+    /// Set redex limit.
+    pub fn redex(mut self, s: usize) -> Self {
+        self.redex = Some(s);
+        self
+    }
 }
 
 impl Core {
@@ -298,7 +690,7 @@ impl Core {
 
     /// Attempt a single-step of VM, under some limits.
     pub fn step(&mut self, limits: &Limits) -> Result<Step, Interruption> {
-        match crate::vm::active_step(self, limits) {
+        match crate::vm_step::active_step(self, limits) {
             Ok(Step {}) => Ok(Step {}),
             Err(Interruption::Send(am, inst, v)) => self.send(limits, am, inst, v),
             Err(Interruption::Response(r)) => self.response(limits, r),
@@ -355,7 +747,7 @@ impl Core {
         });
         let actor = self.actors.map.get_mut(&am.actor).unwrap();
         actor.active = Some(activation);
-        crate::vm::call_function_def(self, actor_env, &f, inst, v)
+        crate::vm_step::call_function_def(self, actor_env, &f, inst, v)
     }
 
     fn response(&mut self, _limits: &Limits, r: Response) -> Result<Step, Interruption> {
@@ -440,7 +832,7 @@ impl Core {
         let source = self.agent.active.cont_source.clone(); // to do -- use prog source
         self.assert_idle_agent()
             .map_err(Interruption::EvalInitError)?;
-        crate::vm::exp_conts_(
+        crate::vm_step::exp_conts_(
             self,
             source.clone(),
             FrameCont::Block,
