+++ crates/motoko/src/lib/vm_ops.rs
@@ -1,16 +1,7 @@
 use crate::ast::{BinOp, PrimType, RelOp, UnOp};
+use crate::value::{Value, Value_};
 use crate::vm_types::Interruption;
-//use crate::ast_traversal::ToNode;
-//use crate::shared::{FastClone, Share};
-use crate::value::{
-    //    ActorId, ActorMethod, Closed, ClosedFunction, CollectionFunction, FastRandIter,
-    //    FastRandIterFunction, HashMapFunction, PrimFunction,
-    Value,
-    //    ValueError,
-    Value_,
-};
 use num_bigint::{BigUint, ToBigInt};
-//use num_traits::ToPrimitive;
 
 use crate::{nyi, type_mismatch};
